requirements.txt
# Core dependencies for our simplified terminal assistant
fastapi==0.104.1
uvicorn==0.24.0

# HTTP client and data validation
httpx==0.25.2
pydantic==2.5.0

# Environment and configuration
python-dotenv==1.0.0

# CLI and terminal interface
click==8.1.7
rich==13.7.0
typer==0.9.0

# OpenAI for LLM integration
openai==1.3.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Cross-platform support
psutil==5.9.6

# JSON-RPC for communication protocol
jsonrpc-base==2.2.0
jsonrpc-async==2.1.2 